+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/usgs/pubs-services/compare/pubs-services-1.0.0...master)
+### Changed
+- Source of IPDS data for dissemination step.
 
 ## [1.0.0](https://github.com/usgs/pubs-services/compare/pubs-services-0.9.4...pubs-services-1.0.0)
 ### Changed